recruitment_dashboard_gui/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openpyxl==3.1.2
pillow==10.2.0
streamlit-option-menu==0.3.12
extra-streamlit-components==0.1.70
python-dotenv==1.0.1
